requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
boto==2.49.0
boto3==1.10.45
botocore==1.13.45
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.3.0
colorama==0.4.3
colorpy==0.1.1
coverage==5.0
cycler==0.10.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
future==0.18.2
gensim==3.8.1
gym==0.17.1
h5py==2.10.0
hue==0.1.0
idna==2.8
jmespath==0.9.4
joblib==0.14.1
jodel-api==1.2.11
Js2Py==0.66
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
mock==3.0.5
more-itertools==8.0.2
mss==5.0.0
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
packaging==19.2
pandas==0.25.3
pathtools==0.1.2
phue==1.1
pipwin==0.4.6
plotly==4.4.1
pluggy==0.13.1
protobuf==3.11.2
psycopg2==2.8.4
py==1.8.0
PyAudio==0.2.11
pycparser==2.19
pyglet==1.5.0
pyjsparser==2.7.1
pyparsing==2.4.5
PyPrind==2.11.2
pyqtgraph==0.10.0
PySocks==1.7.1
pystache==0.5.4
pytest==5.3.2
pytest-cov==2.8.1
python-dateutil==2.8.1
python-twitter==3.5
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
robobrowser==0.5.3
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.3.3
simplejson==3.17.0
six==1.13.0
sklearn==0.0
smart-open==1.9.0
sounddevice==0.3.14
soupsieve==1.9.5
spotipy==2.4.4
sqlparse==0.3.0
static==1.1.1
tweepy==3.8.0
twitter==1.18.0
tzlocal==2.0.0
urllib3==1.25.7
varint==1.0.2
watchdog==0.10.2
wcwidth==0.1.7
Werkzeug==0.16.0
